Restore link to Qt::Core for non-Windows documentation-only builds
The link was removed to work around a build issue on Windows, where we
cannot link Qt::Core against interface libraries with CMake < 3.19.

However, the release documentation is built on Linux, and QDoc requires
a valid include path for QtCore in order to parse the documentation
correctly. Restore the link for non-Windows builds.

Pick-to: 6.8 6.7 6.5
Task-number: QTBUG-111760
Change-Id: Idc53bfdc4e9f6123727eb70795b70721acc171d9
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>